refactor: split event impl and move types
