+++ src/types.ts
@@ -1,5 +1,5 @@
 import type http from 'http'
-import type { H3Event, H3EventContext } from './event'
+import type { H3Event } from './event'
 
 interface CompatibilityRequestProps {
   event: H3Event
@@ -28,3 +28,18 @@ export type Encoding = false | 'ascii' | 'utf8' | 'utf-8' | 'utf16le' | 'ucs2' |
 
 // https://www.rfc-editor.org/rfc/rfc7231#section-4.1
 export type HTTPMethod = 'GET' | 'HEAD' | 'PATCH' | 'POST' | 'PUT' | 'DELETE' | 'CONNECT' | 'OPTIONS' | 'TRACE'
+
+export interface H3EventContext extends Record<string, any> {}
+
+export type CompatibilityEvent = H3Event | IncomingMessage
+
+export type HandlerResponse<T = any> = T | Promise<T>
+
+export interface EventHandler<T = any> {
+  '__is_handler__'?: true
+  (event: H3Event): HandlerResponse<T>
+}
+
+export type LazyEventHandler = () => EventHandler | Promise<EventHandler>
+
+export type CompatibilityEventHandler = EventHandler | Handler | Middleware
